+++ example/lib/pages/picture_in_picture.dart
@@ -1,3 +1,5 @@
+import 'dart:io' show Platform;
+
 import 'package:audio_session/audio_session.dart';
 import 'package:bitmovin_player/bitmovin_player.dart';
 import 'package:bitmovin_player_example/env/env.dart';
@@ -7,6 +9,7 @@ import 'package:logger/logger.dart';
 
 class PictureInPicture extends StatefulWidget {
   const PictureInPicture({super.key});
+
   static String routeName = 'PictureInPicture';
 
   @override
@@ -34,6 +37,21 @@ class _PictureInPictureState extends State<PictureInPicture> {
     ),
   );
   final _eventsKey = GlobalKey<EventsState>();
+  bool _isInPictureInPicture = false;
+
+  void _onPictureInPictureEnterEvent(Event event) {
+    _onEvent(event);
+    setState(() {
+      _isInPictureInPicture = true;
+    });
+  }
+
+  void _onPictureInPictureExitEvent(Event event) {
+    _onEvent(event);
+    setState(() {
+      _isInPictureInPicture = false;
+    });
+  }
 
   void _onEvent(Event event) {
     final eventName = '${event.runtimeType}';
@@ -60,8 +78,25 @@ class _PictureInPictureState extends State<PictureInPicture> {
     super.dispose();
   }
 
+  // Since PiP on Android is basically just the whole activity fitted in a small
+  // floating window, we don't want to display the whole scaffold
+  bool get renderOnlyPlayerView => Platform.isAndroid && _isInPictureInPicture;
+
   @override
   Widget build(BuildContext context) {
+    final playerView = PlayerView(
+      player: _player,
+      key: _playerViewKey,
+      playerViewConfig: _playerViewConfig,
+      onPictureInPictureEnter: _onPictureInPictureEnterEvent,
+      onPictureInPictureEntered: _onEvent,
+      onPictureInPictureExit: _onPictureInPictureExitEvent,
+      onPictureInPictureExited: _onEvent,
+    );
+    if (renderOnlyPlayerView) {
+      return playerView;
+    }
+
     return Scaffold(
       appBar: AppBar(
         title: const Text('Picture-in-Picture'),
@@ -70,15 +105,7 @@ class _PictureInPictureState extends State<PictureInPicture> {
         children: [
           AspectRatio(
             aspectRatio: 16 / 9,
-            child: PlayerView(
-              player: _player,
-              key: _playerViewKey,
-              playerViewConfig: _playerViewConfig,
-              onPictureInPictureEnter: _onEvent,
-              onPictureInPictureEntered: _onEvent,
-              onPictureInPictureExit: _onEvent,
-              onPictureInPictureExited: _onEvent,
-            ),
+            child: playerView,
           ),
           Row(
             children: [